chore: Migrate containeranalysis synth.py to bazel
As suggested in https://github.com/googleapis/java-talent/pull/100, this PR only updates the synth.py itself to use bazel. The updatged generated code will be published separately by yoshi, because it includes breaking changes in resource names, caused not by the bazel migration but by the chnages in the generator itself.